Organize Windows transport by role
