+++ litebox_broker_transport_windows_userland/src/host.rs
@@ -1,29 +1,46 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
+//! Host (broker-side) endpoints of a Windows broker association.
+
 use std::io::{Error, ErrorKind, Result as IoResult};
 use std::sync::{Arc, Mutex};
 use std::thread;
 use std::time::Instant;
 
-use litebox_broker_protocol::message::{BrokerNotification, BrokerRequest, BrokerResponse};
-use litebox_broker_protocol::wire::{
-    WireError, decode_request, encode_notification, encode_response,
+use litebox_broker_protocol::message::{
+    BrokerHandshakeRequest, BrokerHandshakeResponse, BrokerNotification, BrokerRequest,
+    BrokerResponse,
 };
-use litebox_broker_transport::channel::{HostNotificationChannel, HostReceive};
+use litebox_broker_protocol::wire::{
+    WireError, decode_handshake_request, decode_request, encode_handshake_response,
+    encode_notification, encode_response,
+};
+use litebox_broker_transport::channel::{
+    HostNotificationChannel, HostReceive, HostSetupChannel, PeerCredential,
+};
 use litebox_broker_transport::control_ring::{
     CONTROL_RING_READY, ControlRing, ControlRingConsumer, ControlRingProducer,
     ControlRingReadError, ControlRingReadStatus, ControlRingWakeHandle, ControlRingWriteStatus,
 };
+use windows_sys::Win32::Foundation::HANDLE;
+use windows_sys::Win32::System::Pipes::GetNamedPipeClientProcessId;
 
-use crate::named_pipe::WindowsNamedPipeStream;
+use crate::control_ring::PipeLiveness;
+use crate::named_pipe::{TRANSFER_FRAME_TAG, WindowsNamedPipeStream};
 use crate::setup::{
     copy_io_error, invalid_data, read_frame, read_pipe_until_cancelled, ring_error, wire_error,
     write_frame,
 };
-use crate::shared_memory::WindowsSharedMemory;
+use crate::shared_memory::{TransferredSharedMemory, WindowsSharedMemory};
 
-use super::{PipeLiveness, file_handle};
+/// Host-side broker setup channel over a Windows named pipe.
+pub struct WindowsNamedPipeHostSetupChannel {
+    stream: WindowsNamedPipeStream,
+    peer_credential: PeerCredential,
+    setup_deadline: Option<Instant>,
+    negotiated: bool,
+}
 
 /// Request-reading endpoint of an active host control-ring association.
 pub struct WindowsControlRingHostRequestSource {
@@ -63,7 +80,97 @@ enum HostAssociationStatus {
     Failed(Arc<Error>),
 }
 
-pub(crate) fn activate_host(
+impl WindowsNamedPipeHostSetupChannel {
+    /// Creates a host setup channel after the deployment has authenticated the client.
+    pub const fn from_host_guaranteed(
+        stream: WindowsNamedPipeStream,
+        setup_deadline: Instant,
+    ) -> Self {
+        Self {
+            stream,
+            peer_credential: PeerCredential::HostGuaranteed,
+            setup_deadline: Some(setup_deadline),
+            negotiated: false,
+        }
+    }
+
+    /// Duplicates one shared-memory object into the runner and sends its handle values.
+    pub fn send_shared_memory(
+        &mut self,
+        memory: &WindowsSharedMemory,
+        runner_process: HANDLE,
+    ) -> IoResult<()> {
+        let transfer = memory.duplicate_to_process(runner_process)?;
+        write_frame(
+            file_handle(&self.stream),
+            &encode_transfer(&transfer)?,
+            self.setup_deadline,
+        )
+    }
+
+    /// Activates host request, response, and notification control-ring endpoints.
+    pub fn into_active(
+        self,
+        ring: ControlRing<WindowsSharedMemory>,
+    ) -> IoResult<(
+        WindowsControlRingHostRequestSource,
+        WindowsControlRingHostResponseSink,
+        WindowsControlRingHostNotificationChannel,
+        WindowsControlRingHostShutdown,
+    )> {
+        activate_host(self.stream, self.negotiated, self.setup_deadline, ring)
+    }
+}
+
+impl HostSetupChannel for WindowsNamedPipeHostSetupChannel {
+    type Error = Error;
+
+    fn peer_credential(&self) -> IoResult<PeerCredential> {
+        Ok(self.peer_credential)
+    }
+
+    fn recv_handshake_request(&mut self) -> IoResult<HostReceive<BrokerHandshakeRequest>> {
+        let Some(frame) = read_frame(file_handle(&self.stream), self.setup_deadline)? else {
+            return Ok(HostReceive::PeerClosed);
+        };
+        match decode_handshake_request(&frame) {
+            Ok(request) => Ok(HostReceive::Message(request)),
+            Err(WireError::WrongMessagePhase) => Ok(HostReceive::ProtocolViolation),
+            Err(error) => Err(wire_error(error)),
+        }
+    }
+
+    fn send_handshake_response(&mut self, response: &BrokerHandshakeResponse) -> IoResult<()> {
+        write_frame(
+            file_handle(&self.stream),
+            &encode_handshake_response(response.clone()),
+            self.setup_deadline,
+        )?;
+        self.negotiated = matches!(response, BrokerHandshakeResponse::Negotiated { .. });
+        Ok(())
+    }
+}
+
+/// Validates that a connected named-pipe client belongs to `expected_process_id`.
+pub fn validate_client_process(
+    stream: &WindowsNamedPipeStream,
+    expected_process_id: u32,
+) -> IoResult<()> {
+    let mut process_id = 0;
+    // SAFETY: `stream` is a connected named pipe and `process_id` is writable.
+    if unsafe { GetNamedPipeClientProcessId(file_handle(stream), &raw mut process_id) } == 0 {
+        return Err(Error::last_os_error());
+    }
+    if process_id != expected_process_id {
+        return Err(Error::new(
+            ErrorKind::PermissionDenied,
+            "named-pipe client is not the expected runner process",
+        ));
+    }
+    Ok(())
+}
+
+fn activate_host(
     stream: WindowsNamedPipeStream,
     negotiated: bool,
     setup_deadline: Option<Instant>,
@@ -357,7 +464,7 @@ fn monitor_host_pipe(stream: &WindowsNamedPipeStream, association: &HostRingAsso
     match read_pipe_until_cancelled(
         file_handle(stream),
         &mut byte,
-        association.liveness.shutdown_event.handle(),
+        association.liveness.shutdown_handle(),
     ) {
         Ok(0) => association.peer_closed(),
         Ok(_) => {
@@ -370,4 +477,31 @@ fn monitor_host_pipe(stream: &WindowsNamedPipeStream, association: &HostRingAsso
             let _ = association.fail(error);
         }
     }
+}
+
+fn encode_transfer(transfer: &TransferredSharedMemory) -> IoResult<Vec<u8>> {
+    let mut frame = Vec::with_capacity(13 + transfer.handles.len() * 8);
+    frame.push(TRANSFER_FRAME_TAG);
+    frame.extend_from_slice(
+        &u64::try_from(transfer.length)
+            .map_err(|_| invalid_data("shared-memory length is too large"))?
+            .to_le_bytes(),
+    );
+    frame.extend_from_slice(
+        &u32::try_from(transfer.handles.len())
+            .map_err(|_| invalid_data("too many transferred handles"))?
+            .to_le_bytes(),
+    );
+    for handle in &transfer.handles {
+        frame.extend_from_slice(
+            &u64::try_from(*handle)
+                .map_err(|_| invalid_data("transferred handle is too large"))?
+                .to_le_bytes(),
+        );
+    }
+    Ok(frame)
+}
+
+fn file_handle(stream: &WindowsNamedPipeStream) -> HANDLE {
+    stream.handle()
 }
