+++ litebox_broker_transport_windows_userland/src/lib.rs
@@ -9,8 +9,9 @@
 #![cfg(all(windows, target_arch = "x86_64"))]
 
 pub mod control_ring;
+mod host;
+mod local;
+pub mod named_pipe;
 mod pending_calls;
 mod setup;
 pub mod shared_memory;
-
-pub mod named_pipe;
