+++ litebox_broker_transport_windows_userland/src/local.rs
@@ -1,28 +1,51 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
+//! Local (runner-side) endpoints of a Windows broker association.
+
+use std::ffi::OsStr;
+use std::fs::OpenOptions;
 use std::io::{Error, ErrorKind, Result as IoResult};
+use std::os::windows::fs::OpenOptionsExt;
 use std::sync::{Arc, Mutex};
 use std::thread;
-use std::time::Instant;
+use std::time::{Duration, Instant};
 
-use litebox_broker_protocol::message::{BrokerNotification, BrokerRequest, BrokerResponse};
-use litebox_broker_protocol::wire::{decode_notification, decode_response, encode_request};
-use litebox_broker_transport::channel::{LocalCallChannel, LocalNotificationChannel};
+use litebox_broker_protocol::message::{
+    BrokerHandshakeRequest, BrokerHandshakeResponse, BrokerNotification, BrokerRequest,
+    BrokerResponse,
+};
+use litebox_broker_protocol::wire::{
+    decode_handshake_response, decode_notification, decode_response, encode_handshake_request,
+    encode_request,
+};
+use litebox_broker_transport::channel::{
+    LocalCallChannel, LocalNotificationChannel, LocalSetupChannel,
+};
 use litebox_broker_transport::control_ring::{
     CONTROL_RING_READY, ControlRing, ControlRingConsumer, ControlRingProducer,
     ControlRingReadError, ControlRingReadStatus, ControlRingWakeHandle, ControlRingWriteStatus,
 };
+use windows_sys::Win32::Foundation::HANDLE;
+use windows_sys::Win32::Storage::FileSystem::FILE_FLAG_OVERLAPPED;
 
-use crate::named_pipe::WindowsNamedPipeStream;
+use crate::control_ring::PipeLiveness;
+use crate::named_pipe::{TRANSFER_FRAME_TAG, WindowsNamedPipeStream};
 use crate::pending_calls::{PendingCalls, pending_calls_error};
 use crate::setup::{
     copy_io_error, invalid_data, read_frame, read_pipe_until_cancelled, ring_error, wire_error,
     write_frame,
 };
-use crate::shared_memory::WindowsSharedMemory;
+use crate::shared_memory::{TransferredSharedMemory, WindowsSharedMemory};
 
-use super::{PipeLiveness, file_handle};
+const CONNECT_RETRY_DELAY: Duration = Duration::from_millis(10);
+
+/// Local-side broker setup channel over a Windows named pipe.
+pub struct WindowsNamedPipeLocalSetupChannel {
+    pub(crate) stream: WindowsNamedPipeStream,
+    setup_deadline: Option<Instant>,
+    negotiated: bool,
+}
 
 /// Call-issuing endpoint of an active local control-ring association.
 pub struct WindowsControlRingLocalCallChannel {
@@ -44,7 +67,103 @@ struct LocalRingAssociation {
     notification_wake: ControlRingWakeHandle<WindowsSharedMemory>,
 }
 
-pub(crate) fn activate_local(
+impl WindowsNamedPipeLocalSetupChannel {
+    /// Connects to a broker setup pipe until `deadline` expires.
+    pub fn connect_with_setup_deadline(name: &OsStr, deadline: Instant) -> IoResult<Self> {
+        loop {
+            match OpenOptions::new()
+                .read(true)
+                .write(true)
+                .custom_flags(FILE_FLAG_OVERLAPPED)
+                .open(name)
+            {
+                Ok(stream) => {
+                    return Ok(Self {
+                        stream: WindowsNamedPipeStream::from_file(stream),
+                        setup_deadline: Some(deadline),
+                        negotiated: false,
+                    });
+                }
+                Err(error) if Instant::now() < deadline => {
+                    std::thread::sleep(
+                        CONNECT_RETRY_DELAY.min(deadline.saturating_duration_since(Instant::now())),
+                    );
+                    let _ = error;
+                }
+                Err(error) => return Err(error),
+            }
+        }
+    }
+
+    /// Receives and maps one shared-memory object duplicated into this process.
+    pub fn receive_shared_memory(
+        &mut self,
+        expected_length: usize,
+    ) -> IoResult<WindowsSharedMemory> {
+        let frame =
+            read_frame(file_handle(&self.stream), self.setup_deadline)?.ok_or_else(|| {
+                Error::new(
+                    ErrorKind::UnexpectedEof,
+                    "broker closed before transferring shared memory",
+                )
+            })?;
+        let transfer = decode_transfer(&frame)?;
+        if transfer.length != expected_length {
+            return Err(invalid_data("transferred shared-memory length mismatch"));
+        }
+        // SAFETY: The authenticated broker duplicated these handles into this process and encoded
+        // their target-process values in the setup frame.
+        unsafe { WindowsSharedMemory::from_transferred(transfer) }
+    }
+
+    /// Receives and maps the shared control ring duplicated into this process.
+    pub fn receive_control_ring(&mut self) -> IoResult<WindowsSharedMemory> {
+        let frame =
+            read_frame(file_handle(&self.stream), self.setup_deadline)?.ok_or_else(|| {
+                Error::new(
+                    ErrorKind::UnexpectedEof,
+                    "broker closed before transferring control-ring memory",
+                )
+            })?;
+        let transfer = decode_transfer(&frame)?;
+        // SAFETY: The authenticated broker duplicated these handles into this process and encoded
+        // their target-process values in the setup frame.
+        unsafe { WindowsSharedMemory::control_ring_from_transferred(transfer) }
+    }
+
+    /// Activates calls and notifications over the transferred shared control ring.
+    pub fn into_active(
+        self,
+        ring: ControlRing<WindowsSharedMemory>,
+    ) -> IoResult<(
+        WindowsControlRingLocalCallChannel,
+        WindowsControlRingLocalNotificationChannel,
+    )> {
+        activate_local(self.stream, self.negotiated, self.setup_deadline, ring)
+    }
+}
+
+impl LocalSetupChannel for WindowsNamedPipeLocalSetupChannel {
+    type Error = Error;
+
+    fn send_handshake_request(&mut self, request: &BrokerHandshakeRequest) -> IoResult<()> {
+        write_frame(
+            file_handle(&self.stream),
+            &encode_handshake_request(request.clone()),
+            self.setup_deadline,
+        )
+    }
+
+    fn recv_handshake_response(&mut self) -> IoResult<Option<BrokerHandshakeResponse>> {
+        let response = read_frame(file_handle(&self.stream), self.setup_deadline)?
+            .map(|frame| decode_handshake_response(&frame).map_err(wire_error))
+            .transpose()?;
+        self.negotiated = matches!(response, Some(BrokerHandshakeResponse::Negotiated { .. }));
+        Ok(response)
+    }
+}
+
+fn activate_local(
     stream: WindowsNamedPipeStream,
     negotiated: bool,
     setup_deadline: Option<Instant>,
@@ -276,7 +395,7 @@ fn monitor_local_pipe(stream: &WindowsNamedPipeStream, association: &LocalRingAs
     let error = match read_pipe_until_cancelled(
         file_handle(stream),
         &mut byte,
-        association.liveness.shutdown_event.handle(),
+        association.liveness.shutdown_handle(),
     ) {
         Ok(0) => Error::new(
             ErrorKind::UnexpectedEof,
@@ -287,4 +406,34 @@ fn monitor_local_pipe(stream: &WindowsNamedPipeStream, association: &LocalRingAs
         Err(error) => error,
     };
     let _ = association.fail(error);
+}
+
+fn decode_transfer(frame: &[u8]) -> IoResult<TransferredSharedMemory> {
+    if frame.len() < 13 || frame[0] != TRANSFER_FRAME_TAG {
+        return Err(invalid_data("invalid shared-memory transfer frame"));
+    }
+    let length = usize::try_from(u64::from_le_bytes(frame[1..9].try_into().unwrap()))
+        .map_err(|_| invalid_data("transferred shared-memory length is too large"))?;
+    let count = usize::try_from(u32::from_le_bytes(frame[9..13].try_into().unwrap()))
+        .expect("u32 handle count fits usize");
+    if frame.len()
+        != 13
+            + count
+                .checked_mul(8)
+                .ok_or_else(|| invalid_data("handle count overflow"))?
+    {
+        return Err(invalid_data("invalid shared-memory transfer frame length"));
+    }
+    let handles = frame[13..]
+        .chunks_exact(8)
+        .map(|bytes| {
+            usize::try_from(u64::from_le_bytes(bytes.try_into().unwrap()))
+                .map_err(|_| invalid_data("transferred handle is too large"))
+        })
+        .collect::<IoResult<Vec<_>>>()?;
+    Ok(TransferredSharedMemory { length, handles })
+}
+
+fn file_handle(stream: &WindowsNamedPipeStream) -> HANDLE {
+    stream.handle()
 }
